GSYE-333: Moved ForwardLiveEvents from the forward strategy module to its own module.
